Rename package sleeper.systemtest.drivers.statestore
